% File src/library/splines/man/splines-package.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Development Team
% Distributed under GPL 2 or later

\name{splines-package}
\alias{splines-package}
\alias{splines}
\docType{package}
\title{
Regression Spline Functions and Classes
}
\description{
Regression spline functions and classes.
}
\details{
This package provides functions for working with regression
splines using the \link[splines:bs]{B-spline} basis, the
\link[splines:ns]{natural cubic spline} basis.  

For a complete list of functions, use \code{library(help="splines")}.

}
\author{
Douglas M. Bates \email{bates@stat.wisc.edu} and William N. Venables
\email{Bill.Venables@cmis.csiro.au}

Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
\keyword{ models }
