\name{splines-package}
\alias{splines-package}
\alias{splines}
\docType{package}
\title{
Regression Spline Functions and Classes
}
\description{
Regression spline functions and classes.
}
\details{
This package provides functions for working with regression
splines using the \link{splines:bs}{B-spline} basis, the
\link{splines:ns}{natural cubic spline} basis.  

For a complete
list of functions, use \code{library(help="splines")}.

}
\author{
Douglas M. Bates \email{bates@stat.wisc.edu} and William N. Venables
\email{Bill.Venables@cmis.csiro.au}

Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
\keyword{ models }
