% File src/library/splines/man/splines-package.Rd
% Part of the R package, http://www.R-project.org
% Copyright 1995-2007 R Core Team
% Distributed under GPL 2 or later

\name{splines-package}
\alias{splines-package}
\alias{splines}
\docType{package}
\title{
Regression Spline Functions and Classes
}
\description{
Regression spline functions and classes.
}
\details{
This package provides functions for working with regression
splines using the B-spline basis, \code{\link{bs}}, and the
natural cubic spline basis, \code{\link{ns}}.

For a complete list of functions, use \code{library(help = "splines")}.

}
\author{
Douglas M. Bates \email{bates@stat.wisc.edu} and William N. Venables
\email{Bill.Venables@csiro.au}

Maintainer: R Core Team \email{R-core@r-project.org}
}
\keyword{ package }
\keyword{ models }
